Extract quota registration
